feat: split tools.ts into modular builtin tools and separate bridge subsystem
